Cleanup, Refactoring, DB Tests, Basic Crate Structure
