+++ Cargo.toml
@@ -1,17 +1,19 @@
 [workspace.package]
-version = "0.7.8"
+version = "0.7.9"
 edition = "2024"
 license = "MIT"
 
 [workspace]
 resolver = "3"
 
-members = ["crates/bcr-wallet-core", "crates/bcr-wallet-cli"]
+members = ["crates/bcr-wallet-core", "crates/bcr-wallet-cli", "crates/bcr-wallet-persistence", "crates/bcr-wallet-api"]
 
 [workspace.dependencies]
 anyhow = {version = "1"}
 async-trait = {version = "0.1"}
 bcr-common = {git = "https://github.com/BitcreditProtocol/bcr-common.git", tag = "v0.7.0"}
+bcr-wallet-core = {path = "./crates/bcr-wallet-core"}
+bcr-wallet-persistence = {path = "./crates/bcr-wallet-persistence"}
 bip39 = {version = "2.1", features = ["rand"]}
 bitcoin = {version = "0.32"}
 borsh = {version = "1"}
@@ -22,7 +24,6 @@ mockall = {version  = "0.13"}
 nostr = {version = "0.43"}
 nostr-sdk = {version = "0.43", features = ["nip59"]}
 rand = {version = "0.9"}
-redb = {version = "3.1.0"}
 reqwest = {version = "0.12", default-features = false, features = ["json", "rustls-tls"]}
 secp256k1 = {version = "0.29"}
 serde = {version = "1"}
