+++ crates/bcr-wallet-api/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "bcr-wallet-api"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+bcr-common = {workspace = true}
+bcr-wallet-core = {workspace = true}
+bcr-wallet-persistence = {workspace = true, features = ["db-redb"]}
+bip39.workspace = true
+bitcoin.workspace = true
+borsh.workspace = true
+chrono = {workspace = true}
+futures = {workspace = true}
+nostr = {workspace = true}
+nostr-sdk = {workspace = true, features = ["nip06"]}
+rand = {workspace = true}
+reqwest = {workspace = true}
+secp256k1.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+thiserror.workspace = true
+tokio = {workspace = true}
+tracing.workspace = true
+url = {workspace = true}
+uuid = {workspace = true }
+
+[dev-dependencies]
+bcr-common = {workspace = true, features = ["test-utils"]}
+bcr-wallet-persistence = {workspace = true, features = ["test-utils", "db-redb"]}
+mockall = { workspace = true}
