+++ crates/bcr-wallet-api/src/error.rs
@@ -1,6 +1,6 @@
 use anyhow::Error as AnyError;
 use bcr_common::{
-    cashu::{self, MintUrl, nut02 as cdk02},
+    cashu::{self, MintUrl},
     cdk,
 };
 use bitcoin::hashes::sha256::Hash as Sha256;
@@ -45,42 +45,14 @@ pub enum Error {
     NostrClient(#[from] nostr_sdk::client::Error),
     #[error("serde_json: {0}")]
     SerdeJson(#[from] serde_json::Error),
-    #[error("deserialize ciborium: {0}")]
-    CiboriumDe(#[from] ciborium::de::Error<std::io::Error>),
-    #[error("serialize ciborium: {0}")]
-    CiboriumSer(#[from] ciborium::ser::Error<std::io::Error>),
     #[error("reqwest::Url {0}")]
     Url(#[from] url::ParseError),
     #[error("reqwest::Client {0}")]
     ReqwestClient(#[from] reqwest::Error),
     #[error("insufficient funds")]
     InsufficientFunds,
-    #[error("Database operation error: {0}")]
-    Redb(#[from] redb::Error),
-    #[error("Database error: {0}")]
-    RedbDatabase(#[from] redb::DatabaseError),
-    #[error("Database Transaction error: {0}")]
-    RedbTransaction(#[from] redb::TransactionError),
-    #[error("Database Commit error: {0}")]
-    RedbCommit(#[from] redb::CommitError),
-    #[error("Database Table error: {0}")]
-    RedbTable(#[from] redb::TableError),
-    #[error("Database Storage error: {0}")]
-    RedbStorage(#[from] redb::StorageError),
-    #[error("Database Join error: {0}")]
-    RedbTokioSpawn(#[from] tokio::task::JoinError),
-    #[error("proof in local DB not found: {0}")]
-    ProofNotFound(cashu::PublicKey),
-    #[error("proof not in desired state: {0}")]
-    InvalidProofState(cashu::PublicKey),
-    #[error("counter kid mismatch")]
-    CounterKidMismatch,
-    #[error("counter in local DB not found: {0}")]
-    CounterNotFound(cdk02::Id),
     #[error("There already exists a wallet - delete it to create a new one")]
     WalletAlreadyExists,
-    #[error("wallet id {0} not found")]
-    WalletIdNotFound(String),
     #[error("wallet at idx {0} not found")]
     WalletNotFound(usize),
     #[error("empty token: {0}")]
@@ -105,8 +77,6 @@ pub enum Error {
     NoPrepareRef(uuid::Uuid),
     #[error("inactive keyset {0}")]
     InactiveKeyset(cashu::Id),
-    #[error("transaction not found {0}")]
-    TransactionNotFound(cdk::wallet::types::TransactionId),
     #[error("transaction can't be reclaimed - not outgoing or pending {0}")]
     TransactionCantBeReclaimed(cdk::wallet::types::TransactionId),
     #[error("Mint not supporting debit currency")]
@@ -125,10 +95,6 @@ pub enum Error {
     PaymentExpired,
     #[error("melt op unpaid")]
     MeltUnpaid(String),
-    #[error("melt op not found: {0}")]
-    MeltNotFound(String),
-    #[error("mint op not found: {0}")]
-    MintNotFound(String),
     #[error("mint op failed: {0}")]
     MintingError(String),
     #[error("inter-mint payment not supported yet")]
@@ -155,6 +121,8 @@ pub enum Error {
     InsufficientOnChainMeltAmount(u64),
     #[error("insufficient amount for minting {0}")]
     InsufficientOnChainMintAmount(u64),
+    #[error("Database Error: {0}")]
+    Database(#[from] bcr_wallet_persistence::error::Error),
 
     #[error("internal error: {0}")]
     Internal(String),
