+++ crates/bcr-wallet-cli/Cargo.toml
@@ -7,7 +7,7 @@ license.workspace = true
 [dependencies]
 anyhow.workspace = true
 bcr-common.workspace = true
-bcr-wallet-core = {path = "../bcr-wallet-core"}
+bcr-wallet-api= {path = "../bcr-wallet-api"}
 bip39.workspace = true
 bitcoin.workspace = true
 chrono.workspace = true
@@ -16,7 +16,7 @@ config = {version="0.15"}
 nostr-sdk = {workspace = true, features = ["nip06"]}
 serde.workspace = true
 tokio.workspace = true
-toml = "0.8"
+toml = "0.9"
 tracing-log.workspace = true
 tracing-subscriber.workspace = true
 tracing.workspace = true
