--- crates/bcr-wallet-core/src/job.rs
@@ -1,6 +1,0 @@
-use crate::TStamp;
-
-#[derive(Default, Debug, Clone, serde::Serialize, serde::Deserialize)]
-pub struct JobState {
-    pub last_run: TStamp,
-}
