+++ crates/bcr-wallet-core/src/utils.rs
@@ -1,355 +1,13 @@
-use crate::{
-    MintConnector, TStamp,
-    error::{Error, Result},
-    wallet::SafeMode,
+use bitcoin::{
+    hashes::{Hash, HashEngine, sha256},
+    hex::DisplayHex,
 };
-use bcr_common::{
-    cashu::{self, HTLCWitness, Proof},
-    cdk::{self, Error as CdkError},
-    wire::keys::ProofFingerprint,
-};
-use bitcoin::hashes::sha256::Hash as Sha256;
-use bitcoin::secp256k1::PublicKey;
-use secp256k1::schnorr::Signature;
 
-type CdkResult<T> = std::result::Result<T, cdk::Error>;
-
-pub fn tx_can_be_refreshed(tx: &cdk::wallet::types::Transaction) -> bool {
-    // Only refresh outgoing transactions
-    if matches!(
-        tx.direction,
-        cdk::wallet::types::TransactionDirection::Incoming
-    ) {
-        return false;
-    }
-
-    // Only refresh pending transactions
-    let p_status = crate::types::get_transaction_status(&tx.metadata);
-    if !matches!(p_status, crate::types::TransactionStatus::Pending) {
-        return false;
-    }
-    true
-}
-
-pub fn validate_offline_conditions(
-    wallet_pubkey: PublicKey,
-    conditions: &cashu::Conditions,
-    tstamp: u64,
-) -> CdkResult<u64> {
-    tracing::info!("Verifying spending conditions {:?}", conditions);
-
-    let lock_time = conditions.locktime.ok_or(CdkError::LocktimeNotProvided)?;
-    let num_sigs = conditions.num_sigs.ok_or(CdkError::PubkeyRequired)?;
-    let pubkeys = conditions
-        .pubkeys
-        .as_ref()
-        .ok_or(CdkError::PubkeyRequired)?;
-    let refund_len = conditions
-        .refund_keys
-        .as_ref()
-        .map(|r| r.len())
-        .unwrap_or(0);
-
-    if pubkeys.len() != 1 || num_sigs != 1 {
-        return Err(CdkError::PubkeyRequired);
-    }
-    if refund_len != 0 {
-        return Err(CdkError::InvalidSpendConditions(
-            "Beta proofs refund not allowed".into(),
-        ));
-    }
-    if *pubkeys[0] != wallet_pubkey {
-        return Err(CdkError::InvalidSpendConditions(
-            "Pubkey must be wallet pubkey".into(),
-        ));
-    }
-    if lock_time < tstamp + crate::config::LOCK_REDUCTION_SECONDS_PER_HOP {
-        return Err(CdkError::InvalidSpendConditions(
-            "Lock time too short".into(),
-        ));
-    }
-
-    Ok(lock_time)
-}
-
-pub async fn compel_commitment(
-    inputs: Vec<cashu::Proof>,
-    outputs: Vec<cashu::BlindedMessage>,
-    expiration: chrono::TimeDelta,
-    alpha_pk: secp256k1::PublicKey,
-    client: &dyn MintConnector,
-) -> Result<(
-    Vec<cashu::PublicKey>,
-    Vec<cashu::BlindedMessage>,
-    TStamp,
-    secp256k1::schnorr::Signature,
-)> {
-    let result = client
-        .post_commitment(inputs, outputs, expiration, alpha_pk)
-        .await;
-    match result {
-        Ok(response) => Ok(response),
-        // TODO: protest
-        // Err(Error::BorshSignature(BorshMsgSignatureError::Secp256k1(_))) => {},
-        Err(e) => Err(e),
-    }
-}
-
-pub fn proofs_to_fingerprints(
-    proofs: Vec<Proof>,
-) -> Result<(Vec<ProofFingerprint>, Vec<cashu::secret::Secret>)> {
-    let mut secrets = Vec::with_capacity(proofs.len());
-    let mut fingerprints = Vec::with_capacity(proofs.len());
-
-    for p in proofs.iter() {
-        let dleq = p.dleq.clone().ok_or(Error::MissingDleq)?;
-        secrets.push(p.secret.clone());
-
-        fingerprints.push(ProofFingerprint {
-            amount: p.amount.into(),
-            keyset_id: p.keyset_id,
-            c: p.c,
-            dleq: Some(dleq),
-            y: p.y()?,
-            witness: None,
-        });
-    }
-
-    Ok((fingerprints, secrets))
-}
-
-pub fn sign_htlc_proof(
-    proof: &mut Proof,
-    preimage: &str,
-    wallet_secret: &cashu::SecretKey,
-) -> Result<()> {
-    let msg: Vec<u8> = proof.secret.to_bytes();
-    let signature: Signature = wallet_secret
-        .sign(&msg)
-        .map_err(|err| Error::Internal(format!("signing error: {err}")))?;
-
-    let signatures = vec![signature.to_string()];
-
-    proof.witness = Some(cashu::Witness::HTLCWitness(HTLCWitness {
-        preimage: preimage.to_string(),
-        signatures: Some(signatures),
-    }));
-
-    Ok(())
-}
-
-pub async fn htlc_lock(
-    unit: cashu::CurrencyUnit,
-    tstamp: u64,
-    client: &dyn MintConnector,
-    is_credit: bool,
-    proofs: Vec<cashu::Proof>,
-    hash_lock: Sha256,
-    key_locks: Vec<bitcoin::secp256k1::PublicKey>,
-    wallet_pubkey: bitcoin::secp256k1::PublicKey,
-    safe_mode: SafeMode,
-) -> Result<Vec<cashu::Proof>> {
-    tracing::debug!("HTLC-locking proofs");
-    let amount = proofs
-        .iter()
-        .fold(cashu::Amount::ZERO, |acc, x| acc + x.amount);
-
-    let key_locks: Vec<cashu::PublicKey> = key_locks.into_iter().map(|k| k.into()).collect();
-
-    // total hops * time per hop + 2 hops buffer
-    let lock_time =
-        tstamp + (key_locks.len() as u64 + 2) * crate::config::LOCK_REDUCTION_SECONDS_PER_HOP;
-
-    // fetch keysets infos for the given client
-    let infos = client.get_mint_keysets().await?.keysets;
-
-    let active_keyset_id = if is_credit {
-        proofs.first().ok_or(Error::NoActiveKeyset)?.keyset_id
-    } else {
-        infos
-            .iter()
-            .find(|info| info.active && info.unit == unit)
-            .ok_or(Error::NoActiveKeyset)?
-            .id
-    };
-
-    let n = key_locks.len() as u64;
-    let p2pk = cashu::Conditions::new(
-        Some(lock_time),
-        Some(key_locks),
-        Some(vec![wallet_pubkey.into()]),
-        Some(n),
-        None,
-        Some(1),
-    )?;
-    let htlc = cashu::SpendingConditions::new_htlc_hash(&hash_lock.to_string(), Some(p2pk))?;
-    let split_target = cashu::amount::SplitTarget::None;
-    let premints =
-        cashu::PreMintSecrets::with_conditions(active_keyset_id, amount, &split_target, &htlc)?;
-
-    if let SafeMode::Enabled { expire, alpha_pk } = safe_mode {
-        crate::utils::compel_commitment(
-            proofs.clone(),
-            premints.blinded_messages(),
-            expire,
-            alpha_pk,
-            client,
-        )
-        .await?;
-    }
-    let swap_request = cashu::SwapRequest::new(proofs, premints.blinded_messages());
-    let swap = client.post_swap(swap_request).await?;
-
-    let keyset = client.get_mint_keyset(active_keyset_id).await?;
-    let proofs = crate::pocket::unblind_proofs(&keyset, swap.signatures, premints);
-
-    Ok(proofs)
-}
-
-#[cfg(test)]
-pub mod tests {
-    use super::*;
-    use crate::TStamp;
-    use crate::error::Result;
-    use async_trait::async_trait;
-    use bcr_common::{
-        cashu::{
-            nut02 as cdk02, nut03 as cdk03, nut04 as cdk04, nut05 as cdk05, nut06 as cdk06,
-            nut07 as cdk07, nut09 as cdk09, nut23 as cdk23,
-        },
-        cdk_common::Error as CDKError,
-        wire::{keys as wire_keys, melt as wire_melt, mint as wire_mint},
-    };
-
-    use bcr_common::wire::clowder::{AlphaStateResponse, ConnectedMintResponse};
-    type CdkResult<T> = std::result::Result<T, CDKError>;
-
-    mockall::mock! {
-            pub MintConnector {
-            }
-            impl std::fmt::Debug for MintConnector {
-                fn fmt<'a>(&self, f: &mut std::fmt::Formatter<'a>) -> std::fmt::Result;
-            }
-
-            #[async_trait]
-            impl cdk::wallet::MintConnector for MintConnector {
-            async fn get_mint_keys(&self) -> CdkResult<Vec<cdk02::KeySet>>;
-            async fn get_mint_keyset(&self, keyset_id: cdk02::Id) -> CdkResult<cdk02::KeySet>;
-            async fn get_mint_keysets(&self) -> CdkResult<cdk02::KeysetResponse>;
-            async fn post_mint_quote(
-                &self,
-                request: cdk23::MintQuoteBolt11Request,
-            ) -> CdkResult<cdk23::MintQuoteBolt11Response<String>>;
-            async fn get_mint_quote_status(
-                &self,
-                quote_id: &str,
-            ) -> CdkResult<cdk23::MintQuoteBolt11Response<String>>;
-            async fn post_mint(&self, request: cdk04::MintRequest<String>) -> CdkResult<cdk04::MintResponse>;
-            async fn post_melt_quote(
-                &self,
-                request: cdk23::MeltQuoteBolt11Request,
-            ) -> CdkResult<cdk23::MeltQuoteBolt11Response<String>>;
-            async fn get_melt_quote_status(
-                &self,
-                quote_id: &str,
-            ) -> CdkResult<cdk23::MeltQuoteBolt11Response<String>>;
-            async fn post_melt(
-                &self,
-                request: cdk05::MeltRequest<String>,
-            ) -> CdkResult<cdk23::MeltQuoteBolt11Response<String>>;
-            async fn post_swap(&self, request: cdk03::SwapRequest) -> CdkResult<cdk03::SwapResponse>;
-            async fn get_mint_info(&self) -> CdkResult<cdk06::MintInfo>;
-            async fn post_check_state(
-                &self,
-                request: cdk07::CheckStateRequest,
-            ) -> CdkResult<cdk07::CheckStateResponse>;
-            async fn post_restore(&self, request: cdk09::RestoreRequest) -> CdkResult<cdk09::RestoreResponse>;
-    async fn post_mint_bolt12_quote(
-            &self,
-            request: cashu::MintQuoteBolt12Request,
-        ) -> CdkResult<cashu::MintQuoteBolt12Response<String>>;
-        async fn get_mint_quote_bolt12_status(
-            &self,
-            quote_id: &str,
-        ) -> CdkResult<cashu::MintQuoteBolt12Response<String>>;
-        async fn post_melt_bolt12_quote(
-            &self,
-            request: cashu::MeltQuoteBolt12Request,
-        ) -> CdkResult<cashu::MeltQuoteBolt11Response<String>>;
-        async fn get_melt_bolt12_quote_status(
-            &self,
-            quote_id: &str,
-        ) -> CdkResult<cashu::MeltQuoteBolt11Response<String>>;
-        async fn post_melt_bolt12(
-            &self,
-            request: cashu::MeltRequest<String>,
-        ) -> CdkResult<cashu::MeltQuoteBolt11Response<String>>;
-            }
-        #[async_trait]
-        impl crate::MintConnector for MintConnector {
-        async fn get_clowder_betas(&self) -> CdkResult<Vec<cashu::MintUrl>>;
-        fn mint_url(&self) -> cashu::MintUrl;
-
-        async fn post_online_exchange(
-            &self,
-            alpha_proofs: Vec<cashu::Proof>,
-            exchange_path: Vec<bitcoin::secp256k1::PublicKey>,
-        ) -> CdkResult<Vec<cashu::Proof>>;
-        async fn get_clowder_id(&self) -> CdkResult<bitcoin::secp256k1::PublicKey>;
-        async fn post_clowder_path(
-            &self,
-            origin_mint_url: cashu::MintUrl,
-        ) -> CdkResult<bcr_common::wire::clowder::ConnectedMintsResponse>;
-        async fn get_alpha_keysets(
-            &self,
-            alpha_id: bitcoin::secp256k1::PublicKey,
-        ) -> CdkResult<Vec<cashu::KeySet>>;
-
-        async fn get_alpha_offline(&self, alpha_id: bitcoin::secp256k1::PublicKey) -> CdkResult<bool>;
-        async fn get_alpha_status(&self, alpha_id: bitcoin::secp256k1::PublicKey) -> CdkResult<AlphaStateResponse>;
-        async fn get_alpha_substitute(&self, alpha_id: bitcoin::secp256k1::PublicKey) -> CdkResult<ConnectedMintResponse>;
-
-        async fn post_offline_exchange(
-            &self,
-            proofs: Vec<wire_keys::ProofFingerprint>,
-            locks: Vec<bitcoin::hashes::sha256::Hash>,
-            wallet_pubkey: bitcoin::secp256k1::PublicKey,
-        ) -> CdkResult<Vec<cashu::Proof>>;
-
-        async fn post_commitment(
-            &self,
-            inputs: Vec<cashu::Proof>,
-            outputs: Vec<cashu::BlindedMessage>,
-            expiration: chrono::TimeDelta,
-            alpha_pk: secp256k1::PublicKey,
-        ) -> Result<(
-            Vec<cashu::PublicKey>,
-            Vec<cashu::BlindedMessage>,
-            TStamp,
-            secp256k1::schnorr::Signature,
-        )>;
-        async fn post_melt_quote_onchain(
-            &self,
-            req: wire_melt::MeltQuoteOnchainRequest,
-        ) -> Result<wire_melt::MeltQuoteOnchainResponse>;
-        async fn post_melt_onchain(
-            &self,
-            req: cashu::MeltRequest<String>,
-        ) -> Result<wire_melt::MeltQuoteOnchainResponse>;
-        async fn post_mint_quote_onchain(
-            &self,
-            req: wire_mint::MintQuoteOnchainRequest,
-        ) -> Result<wire_mint::MintQuoteOnchainResponse>;
-
-        async fn get_mint_quote_onchain(
-            &self,
-            quote_id: String,
-        ) -> Result<wire_mint::MintQuoteOnchainResponse>;
-
-        async fn post_mint_onchain(
-            &self,
-            req: cashu::MintRequest<String>,
-        ) -> Result<cashu::MintResponse>;
-        }
-    }
+pub fn build_wallet_id(seed: &[u8; 64]) -> String {
+    let mut hasher = sha256::HashEngine::default();
+    hasher.input(seed);
+    sha256::Hash::from_engine(hasher)
+        .as_byte_array()
+        .as_hex()
+        .to_string()
 }
