+++ crates/bcr-wallet-persistence/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "bcr-wallet-persistence"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[features]
+test-utils = ["mockall"]
+db-redb = ["redb"]
+
+[dependencies]
+async-trait.workspace = true
+bcr-common = {workspace = true}
+bcr-wallet-core = {workspace = true}
+bitcoin.workspace = true
+chrono = {workspace = true}
+ciborium = {workspace = true}
+mockall = {workspace = true, optional = true}
+redb = {version = "3.1.0", optional = true}
+serde.workspace = true
+thiserror.workspace = true
+tokio = {workspace = true}
+tracing.workspace = true
+uuid = {workspace = true }
+
+[dev-dependencies]
+bcr-common = {workspace = true, features = ["test-utils"]}
+tokio = {workspace = true}
+
